requirements.txt
flask
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
requests
python-dotenv